+++ Project_2/backend/src/main/java/com/example/Project_2/repositories/TierRepository.java
@@ -1,8 +1,10 @@
 package com.example.Project_2.repositories;
 
-import com.example.Project_2.models.*;
+import com.example.Project_2.models.Tier;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TierRepository extends JpaRepository<Tier, Integer> {}
+public interface TierRepository extends JpaRepository<Tier, Long> {
+
+}
